+++ CHANGELOG.md
@@ -8,7 +8,8 @@
 * The kwarg `pad_mode="constant"` has been added to `AudioTree.from_file(...)`. Other useful choices include `None` (don't pad), and `wrap` (loop audio).
 * New `audiotree.transforms.Roll` Transform that rolls audio forwards or backwards in time with a padding mode.
 * New `audiotree.transforms.Trim` Transform that trims or pads audio to a constant length with a padding mode.
-* New `AudioTree.split_by_batch` to split an AudioTree into a list of AudioTree.
+* New `AudioTree.mini_batch_list` to split an AudioTree into a list of AudioTree.
+* New `AudioTree.mini_batch` and `AudioTree.unbatch` to add or removing a mini-batch axis (useful for `nnx.scan`, etc.)
 * Remove `audiotree.transforms.ReduceBatchTransform` and replace with `audiotree.transforms.Batch`, which should be used in place of `grain.transforms.Batch`
 
 ## audiotree 0.2.0 (Feb 17, 2025)